requirements.txt
Flask==2.1.2
flask-restx==0.5.1
flask-cors==3.0.10
numpy==1.21.6
elasticsearch==8.2.0
kafka-python==2.0.2
minio==7.1.7
pandas==1.4.2
tensorflow==2.8.0
torch==1.11.0
joblib==1.1.0
lightgbm==3.3.2
gunicorn==20.1.0
gevent==21.12.0
protobuf==3.19
python-dateutil==2.8.2
xgboost==1.0.0
scikit-learn==1.0.1
werkzeug==2.1.2